API reference

DataFramesExt.DataFramesExt — Module

DataFramesExt

Types

Functions

Base.read — Function

read(s::IOStream, DataFrame, colnames = nothing;
	quotes = true, comments = true, comment_char = '#') -> DataFrame

Read a DataFrame from the given I/O stream or file where each line gives one row.

If colnames is not provided (or, is nothing), the first row of data will be read as header. If colnames is the symbol :auto, the column names will be x1, x2, and so on. Otherwise, colnames must be a vector of symbols or strings to specify column names.

If quotes is true, columns enclosed within double-quote (") characters are allowed to contain new lines and column delimiters. Double-quote characters within a quoted field must be escaped with another double-quote.

If comments is true, lines beginning with comment_char and text following comment_char in any line are ignored.

Base.read — Method

read(f::AbstractString, DataFrame, colnames = nothing;
	quotes = true, comments = true, comment_char = '#') -> DataFrame

Read a DataFrame from the given I/O stream or file where each line gives one row.

If colnames is not provided (or, is nothing), the first row of data will be read as header. If colnames is the symbol :auto, the column names will be x1, x2, and so on. Otherwise, colnames must be a vector of symbols or strings to specify column names.

If quotes is true, columns enclosed within double-quote (") characters are allowed to contain new lines and column delimiters. Double-quote characters within a quoted field must be escaped with another double-quote.

If comments is true, lines beginning with comment_char and text following comment_char in any line are ignored.

Base.write — Method

write(f::AbstractString, x::AbstractDataFrame; delim = '\t', header = true) -> Int64

Write a DataFrame as text to the given I/O stream or file, using the given delimiter delim (which defaults to tab, but can be anything printable, typically a character or string).

Return the number of bytes written into the stream or file.

Base.write — Method

write(s::IOStream, x::AbstractDataFrame; delim = '\t', header = true) -> Int64

Write a DataFrame as text to the given I/O stream or file, using the given delimiter delim (which defaults to tab, but can be anything printable, typically a character or string).

Return the number of bytes written into the stream or file.