API reference

FITSIOExt.FITSIOExt — Module

FITSIOExt

FITSIODataFramesExt.FITSIODataFramesExt — Module

FITSIODataFramesExt

Types

FITSIO.EitherTableHDU — Type

EitherTableHDU <- Union{ASCIITableHDU, TableHDU}

FITSIO.HDU — Type

HDU  <- Union{ImageHDU, ASCIITableHDU, TableHDU}

Macros

Functions

Base.get — Method

get(f::FITS, name::AbstractString, default::Integer) -> HDU

See also get.

Base.read — Function

read(t::EitherTableHDU, Vector,
	colnames::AbstractVector{<:SymOrStr} = FITSIO.colnames(t)) -> Vector

Read a Vector of columns from the given table (of type ASCIITableHDU or TableHDU). Each column is of type AbstractVector (more specifically, either Vector or AbstractSlices{S, 1} where S).

See also Slices.

Base.read — Function

read(t::EitherTableHDU, Vector{Array},
	colnames::AbstractVector{<:SymOrStr} = FITSIO.colnames(t)) -> Vector

Read a Vector of columns from the given table (of type ASCIITableHDU or TableHDU). Each column is of type Array.

Base.read — Function

read(t::EitherTableHDU, DataFrame,
	colnames::AbstractVector{<:SymOrStr} = FITSIO.colnames(t)) -> DataFrame

Read a DataFrame from the given table (of type ASCIITableHDU or TableHDU).