API reference

StatisticsExt.StatisticsExt — Module

StatisticsExt

Types

Macros

Functions

Exts.nanmean — Method

nanmean(A::AbstractArray, w::AbstractWeights; dims = :)

Compute the weighted mean of array A with weight vector w, and fallback to the unweighted mean if w is all zero(s) (instead of returning NaN). If dims (of type Int) is provided, compute the mean along dimension dims.

See also mean(::AbstractArray), mean(::AbstractArray, ::AbstractWeights).