Exts.jl

Usage

pkg> registry add https://github.com/0h7z/0hjl.git
pkg> add Exts

julia> using Exts

API reference

Types

Exts.Datum — Type

Datum{T} -> Union{T, Missing}
datum(T::Type...) -> Datum{Union{T...}}

Exts.IntOrStr — Type

IntOrStr -> Union{AbstractString, Integer}

Exts.Maybe — Type

Maybe{T} -> Union{T, Nothing}
maybe(T::Type...) -> Maybe{Union{T...}}

Exts.SymOrStr — Type

SymOrStr -> Union{AbstractString, Symbol}

Exts.VecOrTup — Type

VecOrTup -> Union{AbstractVector, Tuple}

Exts.VTuple — Type

VTuple{T} -> Tuple{Vararg{T}}

Functions

Exts.datum — Function

Datum{T} -> Union{T, Missing}
datum(T::Type...) -> Datum{Union{T...}}

Exts.ext — Method

Exts.ext(::Colon) -> VTuple{Pair{Symbol, Maybe{Module}}}

Exts.ext — Method

Exts.ext(x::Symbol) -> Maybe{Module}

Exts.invsqrt — Method

invsqrt(x::T) -> AbstractFloat where T <: Real

Return $\sqrt{x^{-1}}$.

See also sqrt.

Examples

julia> invsqrt(4)
0.5

Exts.isdir — Method

Exts.isdir(path::AbstractString) -> Bool

Return true if path is a directory, false otherwise.

Examples

julia> Exts.isdir(homedir())
true

julia> Exts.isdir("not/a/directory")
false

See also isfile and ispath.

Exts.isdirpath — Method

Exts.isdirpath(path::AbstractString) -> Bool

Determine whether a path refers to a directory (for example, ends with a path separator).

Examples

julia> Exts.isdirpath("/home")
false

julia> Exts.isdirpath("/home/")
true

Exts.lmap — Function

lmap(f, iterators...)

Create a lazy mapping. This is another syntax for writing (f(args...) for args in zip(iterators...)). Equivalent to Iterators.map.

See also map.

Examples

julia> collect(lmap(x -> x^2, 1:3))
3-element Vector{Int64}:
 1
 4
 9

Exts.maybe — Function

Maybe{T} -> Union{T, Nothing}
maybe(T::Type...) -> Maybe{Union{T...}}

Exts.readstr — Method

readstr(x) -> String

Read the entirety of x as a string. Equivalent to read(x, String).

See also read, readchomp.

Exts.stdpath — Method

stdpath(path::AbstractString...; real = false) -> String

Standardize a path (or a set of paths, by joining them together), removing "." and ".." entries and changing path separator to the standard "/".

If path is a directory, the returned path will end with a "/".

If real is true, symbolic links are expanded, however the path must exist in the filesystem. On case-insensitive case-preserving filesystems, the filesystem's stored case for the path is returned.

Examples

julia> stdpath("/home/myuser/../example.jl")
"/home/example.jl"

julia> stdpath("Documents\\Julia\\")
"Documents/Julia/"

Exts.@catch — Macro

@catch expr

Evaluate the expression with a try/catch construct and return the thrown exception. If no error (exception) occurs, return nothing.

See also The-try/catch-statement, try/catch, @trycatch, @try.

Exts.@noinfo — Macro

@noinfo expr

Suppress all lexically-enclosed uses of @debug, @info.

Exts.@nowarn — Macro

@nowarn expr

Suppress all lexically-enclosed uses of @debug, @info, @warn.

Exts.@try — Macro

@try expr default = nothing

Evaluate the expression with a try/catch construct and return the result. If any error (exception) occurs, return default.

See also The-try/catch-statement, try/catch, @trycatch, @catch.

Exts.@trycatch — Macro

@trycatch expr

Evaluate the expression with a try/catch construct and return either the result or the thrown exception, whichever available.

See also The-try/catch-statement, try/catch, @try, @catch.